Optimize project/package dependencies, extract bootstrap logger into separate project, remove left-over ClientCertificateHttpHandler
