+++ src/Bootstrap/test/EmptyAutoConfiguration.Test/EmptyAutoConfigurationTest.cs
@@ -21,6 +21,7 @@ public sealed class EmptyAutoConfigurationTest
             "Steeltoe.Bootstrap.AutoConfiguration.dll",
             "Steeltoe.Common.dll",
             "Steeltoe.Common.Hosting.dll",
+            "Steeltoe.Common.Logging.dll",
 
             "Steeltoe.Bootstrap.EmptyAutoConfiguration.Test.dll",
             "Steeltoe.Common.TestResources.dll"
