+++ src/Common/src/Common/Steeltoe.Common.csproj
@@ -9,17 +9,9 @@
   <Import Project="..\..\..\..\shared.props" />
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.Extensions.Caching.Abstractions" Version="$(FoundationalVersion)" />
-    <PackageReference Include="Microsoft.Extensions.Configuration.Binder" Version="$(FoundationalVersion)" />
-    <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="$(FoundationalVersion)" />
+    <PackageReference Include="Microsoft.Extensions.Configuration" Version="$(FoundationalVersion)" />
     <PackageReference Include="Microsoft.Extensions.Diagnostics.HealthChecks.Abstractions" Version="$(MatchTargetFrameworkVersion)" />
     <PackageReference Include="Microsoft.Extensions.Hosting.Abstractions" Version="$(FoundationalVersion)" />
     <PackageReference Include="Microsoft.Extensions.Options.ConfigurationExtensions" Version="$(FoundationalVersion)" />
-    <PackageReference Include="System.Diagnostics.DiagnosticSource" Version="$(FoundationalVersion)" />
-    <PackageReference Include="System.Reflection.MetadataLoadContext" Version="$(FoundationalVersion)" />
-    <PackageReference Include="System.Text.Json" Version="$(FoundationalVersion)" />
-    <PackageReference Include="Microsoft.Extensions.Configuration.EnvironmentVariables" Version="$(FoundationalVersion)" />
-    <PackageReference Include="Microsoft.Extensions.Configuration.CommandLine" Version="$(FoundationalVersion)" />
-    <PackageReference Include="Microsoft.Extensions.Logging.Console" Version="$(FoundationalVersion)" />
   </ItemGroup>
 </Project>
