+++ src/Common/src/Logging/BootstrapLoggerFactory.cs
