+++ src/Common/src/Logging/IBootstrapLoggerFactory.cs
