+++ src/Common/src/Logging/Steeltoe.Common.Logging.csproj
@@ -1,14 +1,14 @@
 <Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
     <TargetFrameworks>net8.0</TargetFrameworks>
-    <Description>Steeltoe common library for HTTP</Description>
-    <PackageTags>http; discovery; loadbalancing</PackageTags>
+    <Description>Steeltoe library for common logging-related utilities</Description>
+    <PackageTags>logging</PackageTags>
     <IsPackable>true</IsPackable>
   </PropertyGroup>
 
   <Import Project="..\..\..\..\shared.props" />
 
   <ItemGroup>
-    <ProjectReference Include="..\Common\Steeltoe.Common.csproj" />
+    <ProjectReference Include="..\Hosting\Steeltoe.Common.Hosting.csproj" />
   </ItemGroup>
 </Project>
