+++ src/Common/src/Logging/UpgradableBootstrapLoggerFactory.cs
