+++ src/Common/test/Common.Test/Steeltoe.Common.Test.csproj
@@ -8,10 +8,4 @@
   <ItemGroup>
     <ProjectReference Include="..\..\src\Common\Steeltoe.Common.csproj" />
   </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.Extensions.Caching.Memory" Version="$(FoundationalVersion)" />
-    <PackageReference Include="Microsoft.Extensions.Configuration.Json" Version="$(FoundationalVersion)" />
-    <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="$(FoundationalVersion)" />
-  </ItemGroup>
 </Project>
