+++ src/Common/test/Http.Test/Steeltoe.Common.Http.Test.csproj
@@ -8,9 +8,4 @@
   <ItemGroup>
     <ProjectReference Include="..\..\src\Http\Steeltoe.Common.Http.csproj" />
   </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.Extensions.Configuration.Json" Version="$(FoundationalVersion)" />
-    <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="$(FoundationalVersion)" />
-  </ItemGroup>
 </Project>
