+++ src/Configuration/src/CloudFoundry.ServiceBinding/Steeltoe.Configuration.CloudFoundry.ServiceBinding.csproj
@@ -10,7 +10,7 @@
   <Import Project="..\..\..\..\shared.props" />
 
   <ItemGroup>
-    <FrameworkReference Include="Microsoft.AspNetCore.App" />
+    <PackageReference Include="Microsoft.Extensions.Configuration.Json" Version="$(FoundationalVersion)" />
   </ItemGroup>
 
   <ItemGroup>
