+++ src/Configuration/src/Kubernetes.ServiceBinding/Steeltoe.Configuration.Kubernetes.ServiceBinding.csproj
@@ -10,7 +10,7 @@
   <Import Project="..\..\..\..\shared.props" />
 
   <ItemGroup>
-    <FrameworkReference Include="Microsoft.AspNetCore.App" />
+    <PackageReference Include="Microsoft.Extensions.FileProviders.Physical" Version="$(FoundationalVersion)" />
   </ItemGroup>
 
   <ItemGroup>
