+++ src/Configuration/src/RandomValue/Steeltoe.Configuration.RandomValue.csproj
@@ -10,11 +10,6 @@
   <Import Project="..\..\..\..\shared.props" />
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.Extensions.Configuration" Version="$(FoundationalVersion)" />
-    <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="$(FoundationalVersion)" />
-  </ItemGroup>
-
-  <ItemGroup>
     <ProjectReference Include="..\..\..\Common\src\Common\Steeltoe.Common.csproj" />
   </ItemGroup>
 </Project>
