+++ src/Configuration/src/SpringBoot/Steeltoe.Configuration.SpringBoot.csproj
@@ -10,13 +10,8 @@
   <Import Project="..\..\..\..\shared.props" />
 
   <ItemGroup>
-    <FrameworkReference Include="Microsoft.AspNetCore.App" />
-  </ItemGroup>
-
-  <ItemGroup>
     <PackageReference Include="Microsoft.Extensions.Configuration.CommandLine" Version="$(FoundationalVersion)" />
     <PackageReference Include="Microsoft.Extensions.Configuration.Json" Version="$(FoundationalVersion)" />
-    <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="$(FoundationalVersion)" />
   </ItemGroup>
 
   <ItemGroup>
