+++ src/Configuration/test/CloudFoundry.ServiceBinding.Test/Steeltoe.Configuration.CloudFoundry.ServiceBinding.Test.csproj
@@ -9,8 +9,4 @@
   <ItemGroup>
     <ProjectReference Include="..\..\src\CloudFoundry.ServiceBinding\Steeltoe.Configuration.CloudFoundry.ServiceBinding.csproj" />
   </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.Extensions.Options.ConfigurationExtensions" Version="$(FoundationalVersion)" />
-  </ItemGroup>
 </Project>
