+++ src/Configuration/test/CloudFoundry.Test/Steeltoe.Configuration.CloudFoundry.Test.csproj
@@ -9,8 +9,4 @@
   <ItemGroup>
     <ProjectReference Include="..\..\src\CloudFoundry\Steeltoe.Configuration.CloudFoundry.csproj" />
   </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.Extensions.Options.ConfigurationExtensions" Version="$(FoundationalVersion)" />
-  </ItemGroup>
 </Project>
