+++ src/Configuration/test/Kubernetes.ServiceBinding.Test/Steeltoe.Configuration.Kubernetes.ServiceBinding.Test.csproj
@@ -15,8 +15,4 @@
   <ItemGroup>
     <ProjectReference Include="..\..\src\Kubernetes.ServiceBinding\Steeltoe.Configuration.Kubernetes.ServiceBinding.csproj" />
   </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.Extensions.Options.ConfigurationExtensions" Version="$(FoundationalVersion)" />
-  </ItemGroup>
 </Project>
