+++ src/Discovery/src/HttpClients/Steeltoe.Discovery.HttpClients.csproj
@@ -10,7 +10,8 @@
   <Import Project="..\..\..\..\shared.props" />
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.Extensions.Hosting.Abstractions" Version="$(FoundationalVersion)" />
+    <PackageReference Include="Microsoft.Extensions.Caching.Abstractions" Version="$(FoundationalVersion)" />
+    <PackageReference Include="Microsoft.Extensions.Http" Version="$(FoundationalVersion)" />
   </ItemGroup>
 
   <ItemGroup>
