+++ src/Logging/src/DynamicLogger/Steeltoe.Logging.DynamicLogger.csproj
@@ -10,13 +10,6 @@
   <Import Project="..\..\..\..\shared.props" />
 
   <ItemGroup>
-    <FrameworkReference Include="Microsoft.AspNetCore.App" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.Extensions.Hosting" Version="$(FoundationalVersion)" />
-    <PackageReference Include="Microsoft.Extensions.Configuration.Abstractions" Version="$(FoundationalVersion)" />
-    <PackageReference Include="Microsoft.Extensions.Options.ConfigurationExtensions" Version="$(FoundationalVersion)" />
     <PackageReference Include="Microsoft.Extensions.Logging.Console" Version="$(FoundationalVersion)" />
   </ItemGroup>
