+++ src/Logging/src/DynamicSerilog/Steeltoe.Logging.DynamicSerilog.csproj
@@ -10,18 +10,12 @@
   <Import Project="..\..\..\..\shared.props" />
 
   <ItemGroup>
-    <FrameworkReference Include="Microsoft.AspNetCore.App" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.Extensions.Hosting" Version="$(FoundationalVersion)" />
     <PackageReference Include="Serilog.Extensions.Logging" Version="$(SerilogVersion)" />
     <PackageReference Include="Serilog.Settings.Configuration" Version="$(SerilogVersion)" />
     <PackageReference Include="Serilog.Sinks.Console" Version="$(SerilogSinksConsoleVersion)" />
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\..\Common\src\Common\Steeltoe.Common.csproj" />
     <ProjectReference Include="..\..\..\Common\src\Hosting\Steeltoe.Common.Hosting.csproj" />
     <ProjectReference Include="..\Abstractions\Steeltoe.Logging.Abstractions.csproj" />
   </ItemGroup>
