+++ src/Management/src/Abstractions/Steeltoe.Management.Abstractions.csproj
@@ -11,11 +11,6 @@
   <Import Project="..\..\..\..\shared.props" />
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.Extensions.Logging" Version="$(FoundationalVersion)" />
-    <PackageReference Include="System.Diagnostics.DiagnosticSource" Version="$(FoundationalVersion)" />
-  </ItemGroup>
-
-  <ItemGroup>
     <ProjectReference Include="..\..\..\Common\src\Common\Steeltoe.Common.csproj" />
   </ItemGroup>
 </Project>
